# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Archivista API",
        "title": "Archivista API",
        "contact": {
            "name": "Archivista Contributors",
            "url": "https://github.com/ramidij/archivista/issues/new"
        },
        "license": {
            "url": "https://opensource.org/licenses/Apache-2"
        },
        "version": "v1"
    },
    "paths": {
        "/download/{gitoid}": {
            "get": {
                "description": "download an attestation",
                "produces": [
                    "application/json"
                ],
                "summary": "Download",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "gitoid",
                        "name": "gitoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dsse.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "stores an attestation",
                "produces": [
                    "application/json"
                ],
                "summary": "Upload",
                "deprecated": true,
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StoreResponse"
                        }
                    }
                }
            }
        },
        "/v1/artifacts": {
            "get": {
                "description": "retrieves details about all available Artifacts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artifacts"
                ],
                "summary": "List all Artifacts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/artifactstore.Artifact"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/artifacts/{name}": {
            "get": {
                "description": "retrieves details about all available versions of a specified artifact",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artifacts"
                ],
                "summary": "List Artifact Versions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "artifact name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/artifactstore.Version"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/artifacts/{name}/{version}": {
            "get": {
                "description": "retrieves details about a specified version of an artifact",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artifacts"
                ],
                "summary": "Artifact Version Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "artifact name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "version of artifact",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/artifactstore.Version"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "objecpec"
                        }
                    }
                }
            }
        },
        "/v1/download/artifact/{name}/{version}/{distribution}": {
            "get": {
                "description": "downloads a specified distribution of an artifact",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Artifacts"
                ],
                "summary": "Download Artifact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of artifact",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "version of artifact to download",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "distribution of artifact to download",
                        "name": "distribution",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/download/{gitoid}": {
            "get": {
                "description": "download an attestation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attestation"
                ],
                "summary": "Download",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gitoid",
                        "name": "gitoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dsse.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/query": {
            "post": {
                "description": "GraphQL query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graphql"
                ],
                "summary": "Query GraphQL",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/archivista.Resolver"
                        }
                    }
                }
            }
        },
        "/v1/upload": {
            "post": {
                "description": "stores an attestation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attestation"
                ],
                "summary": "Upload",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StoreResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.StoreResponse": {
            "type": "object",
            "properties": {
                "gitoid": {
                    "type": "string"
                }
            }
        },
        "archivista.Resolver": {
            "type": "object"
        },
        "artifactstore.Artifact": {
            "type": "object",
            "properties": {
                "versions": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/artifactstore.Version"
                    }
                }
            }
        },
        "artifactstore.Distribution": {
            "type": "object",
            "properties": {
                "sha256digest": {
                    "type": "string"
                }
            }
        },
        "artifactstore.Version": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "distributions": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/artifactstore.Distribution"
                    }
                }
            }
        },
        "dsse.Envelope": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "payloadType": {
                    "type": "string"
                },
                "signatures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dsse.Signature"
                    }
                }
            }
        },
        "dsse.Signature": {
            "type": "object",
            "properties": {
                "certificate": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "intermediates": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "keyid": {
                    "type": "string"
                },
                "sig": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "timestamps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dsse.SignatureTimestamp"
                    }
                }
            }
        },
        "dsse.SignatureTimestamp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "type": {
                    "$ref": "#/definitions/dsse.SignatureTimestampType"
                }
            }
        },
        "dsse.SignatureTimestampType": {
            "type": "string",
            "enum": [
                "tsp"
            ],
            "x-enum-varnames": [
                "TimestampRFC3161"
            ]
        }
    }
}